Set up the basic notes fragment and text editor
Set up the notes fragment showing created notes as vertical RecyclerView items in reverse chronological order.  Items are fetched by a SQLDatabase. The item views contain small icons as note type indication (text, voice or picture). Clicking on an item will open an editor to edit it - currently only a feature to edit text note items is implemented.

Also add a feature for swiping left to delete an item from the SQLDatabase and thereby from the list view. A toast has been added with the clickable option of undoing the deletion for a short period of time.

Also set up a button to add new notes with a small pop-up menu for note type selection. Currently only the feature of adding text notes is implemented.

Also set up a basic text editor, where the edit text is either empty (using the add note button) or already containing the note text to edit (clicking on an already existing text note item in the list view). The editor contains a vertical menu bar for formatting the text by applying it on either selected text parts  and keeping it for further input or without selection only on further input. The applied formatting option is highlighted in each case.