Implement SwipeToDelete on item view
Replace SwipeToDelete-notification and undo-option on swiped list items' view instead of a separate snackbar on the bottom of the display. In detail: In SwipeToDeleteCallback the snackbar has been removed and onChildDraw a canvas will be created and a notification message gets drawn on it. A drawable for the undo option gets drawn on the canvas as well and a custom OnTouchListener has been placed within its bounds, since the drawn object is not clickable.

Also some small adjustments like replacing a switch-statement with an if-statement or renaming values.

Also removed unnecessary restore-Item method in RecyclerViewAdapter.

Also added a small textview to the note list item views to display its last modification date.

Also fixed a bug with saving create and mod dates for notes: former int variables could not save the current date variable (value has been too long) and had to replaced with variables of type Long.