Setup select and delete multiple note items
Add feature to select listview items in NotesFragment, which highlights them  by backgroundColorChange.
On select at least one item a panel is shown on the bottom (bottom-up animation) and hidden (by bottom-down-animation) whenever selected items are removed or all items are deselected.
The panel contains a ImageButton (delete-icon) to remove all selected items from the database and the recyclerViewList.