Clean up RichTextEditor and bug fix of underline-feature
Adjust apply-cases for the toolbar features for the RichTextEditor in TextNoteEditorFragment: instead of checking backgroundcolor-changes the toolbar icon views are declared selected or not and boolean is passed to RichTextEditor. Only undo and redo are exceptions and only get backgroundcolor-changes on click and switch back after a short time.

Also fix the quote span appearance and small bugs on its appliance.

Also fix underline appliance bug (there were continue statements instead of returns).

Also checkstyled all changed files.

Also fix a bug where applied text formats did get applied to the entire text end instead of the line or point of cursor when nothing was selected (adjust endOfString-variable in onTextChanged).