Bugfix: insert new lines when any text format is applied is causing overlapping text
Fix a bug, which is causing text lines to overlap when a text format is applied and new lines were inserted (break lines). This was caused by the way the span start position was declared: it is now declared in the onTextChanged method as it is one index number lower than the current cursor position, if text is inserted (checked by text length before and after input) instead of saving a position to the variable whenever the format method is called.