Adjust text note editor layout and onClick to new toolbar interface
Setup a new text format button in the header area, which is showing and hiding the unchanged) text format toolbar just below onClick.

When the toolbar is hidden, there is a small arrow "down" below the format indicator icon, whenever it is is shown it is switching to a small arrow "up".