Adjust the BookNotesRecyclerViewAdapter to handle voice note features
Adjust the BookNotesRecyclerViewAdapter to handle voice note features like play/pause and stop and showing an visualizer on playing audio.

Also adjusted the cardView to match the general notes one in NotesFragment.

Also adjust the reverse-chronological order to match the NotesFragment sorting and improve usability.

Also rename inner holder-class to NotesViewHolder and adjust it to handle VoiceNote-view-items.