Update book mod_date with new notes
For the display of recently edited books in the home fragment, it is important that the mod_date of a book is also sensitive to new notes, not only if the metadata of books is changed. Therefore, when saving a note, the mod_date of the book is now updated as well.